backend/requirements.txt
flask==3.0.0
numpy==1.26.2
tensorflow==2.16.1
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
flask-cors==4.0.0
scikit-image==0.22.0
imageio==2.31.1
scipy>=1.11.3
networkx>=3.0
tifffile>=2023.7.10
lazy-loader>=0.3
